🚚 Rename `Manager` to `QueryManager` (#984)
