+++ lamindb/_filter.py
@@ -2,7 +2,7 @@ from typing import Type
 
 from lnschema_core import Registry
 
-from lamindb._queryset import QuerySet
+from lamindb._query_set import QuerySet
 
 
 def filter(Registry: Type[Registry], **expressions) -> QuerySet:
