+++ lamindb/dev/__init__.py
@@ -5,7 +5,7 @@
 
    Registry
    QuerySet
-   Manager
+   QueryManager
    FeatureManager
    datasets
    hashing
@@ -17,8 +17,8 @@
 from lnschema_core.models import Registry
 
 from lamindb._feature_manager import FeatureManager
-from lamindb._manager import Manager
-from lamindb._queryset import QuerySet
+from lamindb._query_manager import QueryManager
+from lamindb._query_set import QuerySet
 
 from .._context import run_context
 from . import datasets  # noqa
